[#146] 데이터 연결
