+++ Projects/Domain/AnimalListDomain/Sources/AnimalListDomainResources.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public enum AnimalListDomainResources {
+    public static let bundle = Bundle.module
+}
