+++ Projects/Feature/AnimalQuizLabFeature/Sources/AQLConstants.swift
@@ -2,6 +2,7 @@ import Foundation
 import GameEntityDomainInterface
 
 struct AQLConstants {
+    static var bundle: Bundle!
     static var gameUseCase: GameListUseCase!
     static var categoryUseCase: CategoryListUseCase!
 }
