+++ Projects/Feature/AnimalQuizLabFeature/Sources/AQLPayload.swift
@@ -1,10 +1,17 @@
+import Foundation
 import GameEntityDomainInterface
 
 public struct AQLPayload {
+    let bundle: Bundle
     let listUseCase: GameListUseCase
     let categoryListUseCase: CategoryListUseCase
     
-    public init(listUseCase: GameListUseCase, categoryListUseCase: CategoryListUseCase) {
+    public init(
+        bundle: Bundle,
+        listUseCase: GameListUseCase,
+        categoryListUseCase: CategoryListUseCase
+    ) {
+        self.bundle = bundle
         self.listUseCase = listUseCase
         self.categoryListUseCase = categoryListUseCase
     }
